requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
Jinja2==2.11.3
joblib==1.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.4
protobuf==3.15.6
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
six==1.15.0
sklearn==0.0
tensorflow-cpu==2.4.1
urllib3==1.26.4
